downloader: handle conferences without end date
Some conferences might not have an end date, but might be disabled in
their config nonetheless. This leads to a situation which causes a null
dereference in the downloader code.

Fix this by not downloading data for conferences in such a state, as it
stands to reason that we don't care about them anymore if we explicitly
disabled them.